requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# LangChain & LangGraph
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.24
langgraph==0.0.26
langsmith==0.0.92

# Azure SDKs
azure-identity==1.15.0
azure-cosmos==4.5.1
azure-search-documents==11.4.0
azure-monitor-opentelemetry==1.2.0
openai==1.12.0

# Data Processing
pandas==2.2.0
numpy==1.26.4

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.3

# Validation & Security
email-validator==2.1.0.post1
python-multipart==0.0.9

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
